cogs.reminders: Add secret reminders
Secret reminders allow a user to hide a reminder from the public.

Secret reminders will not show the description of the reminder if `remind list` is run in a guild. When ran in DMs, the description will show up in the embed.
When it's time to remind the user, the bot will DM the user the reminder.

By default, any reminder is not marked as secret until the user explicitly runs `remind hide`.

To hide a reminder, use `remind hide`. To do the inverse, use `remind show`.